pyproject.toml
[project]
name = "datazoomcamp-mlops-final-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Mangilev",email = "ssmangilev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "xgboost (>=3.0.0,<4.0.0)",
    "apache-airflow (>=3.0.0,<4.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "polars (>=1.29.0,<2.0.0)",
    "hyperopt (>=0.2.7,<0.3.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "setuptools (>=80.4.0,<81.0.0)"
]

[tool.poetry]
packages = [
    {include = "orchestration"}
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.8"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
